pipeline/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-blob==12.8.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cmdstanpy==0.9.68
convertdate==2.3.2
cycler==0.10.0
Cython==0.29.21
ephem==3.7.7.1
grpcio==1.37.1
hijri-converter==2.1.1
holidays==0.11.1
httplib2==0.19.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
libcst==0.3.19
LunarCalendar==0.0.9
matplotlib==3.4.2
munch==2.5.0
mypy-extensions==0.4.3
nltk==3.6.2
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyenchant==3.2.1
PyMeeus==0.5.11
pygeos==0.7.1
pyparsing==2.4.7
pyproj==3.0.1
PySocks==1.7.1
pyspellchecker==0.6.2
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-git==1.2
shapely==1.7.0
six==1.16.0
smart-open==5.0.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.4
wincertstore==0.2
openpyxl==3.0.7
lxml==4.6.3